+++ lib/stagehand.dart
@@ -32,6 +32,7 @@ import 'dart:convert';
 
 import 'src/common.dart';
 import 'src/generators/console_full.dart';
+import 'src/generators/flutter_web_preview.dart';
 import 'src/generators/package_simple.dart';
 import 'src/generators/server_shelf.dart';
 import 'src/generators/web_angular.dart';
@@ -41,6 +42,7 @@ import 'src/generators/web_stagexl.dart';
 /// A curated, prescriptive list of Dart project generators.
 final List<Generator> generators = [
   ConsoleFullGenerator(),
+  FlutterWebPreviewGenerator(),
   PackageSimpleGenerator(),
   ServerShelfGenerator(),
   WebAngularGenerator(),